+++ otlp/queue_manager.go
@@ -27,10 +27,13 @@ import (
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/config"
 	"github.com/lightstep/opentelemetry-prometheus-sidecar/telemetry/doevery"
 	"github.com/pkg/errors"
+	"github.com/prometheus/common/version"
 	promconfig "github.com/prometheus/prometheus/config"
 	"go.opentelemetry.io/otel/metric"
 	metricsService "go.opentelemetry.io/proto/otlp/collector/metrics/v1"
-	metric_pb "go.opentelemetry.io/proto/otlp/metrics/v1"
+	metricspb "go.opentelemetry.io/proto/otlp/metrics/v1"
+	resourcepb "go.opentelemetry.io/proto/otlp/resource/v1"
+	commonpb "go.opentelemetry.io/proto/otlp/common/v1"
 
 	// gRPC Status protobuf types we may want to see.  This type
 	// is not widely used, but is the most standard way to itemize
@@ -50,6 +53,13 @@ const (
 	shardUpdateDuration = 15 * time.Second
 
 	maxErrorDetailStringLen = 512
+)
+
+var (
+	staticInstrumentationLibrary = &commonpb.InstrumentationLibrary{
+		Name:    sidecar.ExportInstrumentationLibrary,
+		Version: version.Version,
+	}
 )
 
 // StorageClient defines an interface for sending a batch of samples to an
@@ -81,6 +91,7 @@ type QueueManager struct {
 	cfg           promconfig.QueueConfig
 	timeout       time.Duration
 	clientFactory StorageClientFactory
+	resource      *resourcepb.Resource
 
 	shardsMtx   sync.RWMutex
 	shards      *shardCollection
@@ -105,7 +116,7 @@ type QueueManager struct {
 }
 
 // NewQueueManager builds a new QueueManager.
-func NewQueueManager(logger log.Logger, cfg promconfig.QueueConfig, timeout time.Duration, clientFactory StorageClientFactory, tailer LogReaderClient) (*QueueManager, error) {
+func NewQueueManager(logger log.Logger, cfg promconfig.QueueConfig, timeout time.Duration, clientFactory StorageClientFactory, tailer LogReaderClient, resource *resourcepb.Resource) (*QueueManager, error) {
 	if logger == nil {
 		logger = log.NewNopLogger()
 	}
@@ -128,6 +139,7 @@ func NewQueueManager(logger log.Logger, cfg promconfig.QueueConfig, timeout time
 		walSize:            newEWMARate(ewmaWeight, shardUpdateDuration),
 		walOffset:          newEWMARate(ewmaWeight, shardUpdateDuration),
 		tailer:             tailer,
+		resource:           resource,
 	}
 	lastSize, err := tailer.Size()
 	if err != nil {
@@ -202,7 +214,7 @@ func NewQueueManager(logger log.Logger, cfg promconfig.QueueConfig, timeout time
 
 // Append queues a sample to be sent to the OpenTelemetry API.
 // Always returns nil.
-func (t *QueueManager) Append(ctx context.Context, hash uint64, sample *metric_pb.ResourceMetrics) error {
+func (t *QueueManager) Append(ctx context.Context, hash uint64, sample *metricspb.Metric) error {
 	t.queueLengthCounter.Add(ctx, 1)
 
 	t.shardsMtx.RLock()
@@ -390,7 +402,7 @@ func (t *QueueManager) reshard(n int) {
 }
 
 type queueEntry struct {
-	sample *metric_pb.ResourceMetrics
+	sample *metricspb.Metric
 }
 
 type shard struct {
@@ -434,7 +446,7 @@ func (s *shardCollection) stop() {
 	}
 }
 
-func (s *shardCollection) enqueue(hash uint64, sample *metric_pb.ResourceMetrics) {
+func (s *shardCollection) enqueue(hash uint64, sample *metricspb.Metric) {
 	s.qm.samplesIn.incr(1)
 	shardIndex := hash % uint64(len(s.shards))
 
@@ -450,7 +462,7 @@ func (s *shardCollection) runShard(i int) {
 	shard := s.shards[i]
 
 	// Send batches of at most MaxSamplesPerSend samples to the remote storage.
-	pendingSamples := make([]*metric_pb.ResourceMetrics, 0, s.qm.cfg.MaxSamplesPerSend)
+	pendingSamples := make([]*metricspb.Metric, 0, s.qm.cfg.MaxSamplesPerSend)
 
 	ctx := context.Background()
 
@@ -503,7 +515,7 @@ func (s *shardCollection) runShard(i int) {
 	}
 }
 
-func (s *shardCollection) sendSamples(client StorageClient, samples []*metric_pb.ResourceMetrics) {
+func (s *shardCollection) sendSamples(client StorageClient, samples []*metricspb.Metric) {
 	begin := time.Now()
 	s.sendSamplesWithBackoff(client, samples)
 
@@ -514,7 +526,7 @@ func (s *shardCollection) sendSamples(client StorageClient, samples []*metric_pb
 }
 
 // sendSamples to the remote storage with backoff for recoverable errors.
-func (s *shardCollection) sendSamplesWithBackoff(client StorageClient, samples []*metric_pb.ResourceMetrics) {
+func (s *shardCollection) sendSamplesWithBackoff(client StorageClient, samples []*metricspb.Metric) {
 	ctx := context.Background()
 	start := time.Now()
 	backoff := s.qm.cfg.MinBackoff
@@ -524,10 +536,22 @@ func (s *shardCollection) sendSamplesWithBackoff(client StorageClient, samples [
 	// repeatedly time out.
 	maxWait := s.qm.timeout * time.Duration(config.DefaultMaxExportAttempts)
 
+	req := &metricsService.ExportMetricsServiceRequest{
+		ResourceMetrics: []*metricspb.ResourceMetrics{
+			{
+				Resource: s.qm.resource,
+				InstrumentationLibraryMetrics: []*metricspb.InstrumentationLibraryMetrics{
+					{
+						InstrumentationLibrary: staticInstrumentationLibrary,
+						Metrics:                samples,
+					},
+				},
+			},
+		},
+	}
+
 	for time.Since(start) < maxWait {
-		err := client.Store(&metricsService.ExportMetricsServiceRequest{
-			ResourceMetrics: samples,
-		})
+		err := client.Store(req)
 
 		if err == nil {
 			s.qm.sendOutcomesCounter.Add(